Fixes descriptors for broad band loaders
It turns out that the C6 product doesn't really have the same QA
for individual narrow bands as for broad bands, so I have had to
hack this somehow. We might need to go back to how the QA for the
vis and NIR bands is extracted from the A1 product, as I haven't
really found much information on how to interpret the data, and
how different it is from C5.